+++ HISTORY.rst
@@ -9,6 +9,8 @@ As of 1.0.0, `semantic versioning <http://semver.org/>`__ is used.
 +++++++++
 released XXXX-XX-XX
 
+- fix mobileclient.get_genres: return a list and handle invalid parent genres.
+
 
 4.0.0
 +++++